Adding the first tests for our ViewModel
